Store private files on Dropbox

+++ setup_links.sh
@@ -1,57 +1,92 @@
 #!/usr/bin/env bash
 
+# ------------------------------------------------------------------------------
 # Settings
-sync_folder="$HOME/Google Drive"
+# ------------------------------------------------------------------------------
+SYNC_FOLDER="$HOME/Dropbox"
 
-echo "Setting links to dotfiles on user home dir: $HOME"
+# ------------------------------------------------------------------------------
+# Utility Funtions
+# ------------------------------------------------------------------------------
+RED="\e[31m"
+REDBOLD="\e[31m\e[1m"
+GREEN="\e[32m"
+YELLOW="\e[33m"
+RESET="\e[0m"
+
+log () {
+    if [ $2 ]; then
+        echo `printf "$2$1 $RESET"`
+    else
+        echo `printf "$RESET$1 $RESET"`
+    fi
+}
 
 create_link() {
   origin=$1
   dest=$2
-  echo Linking origin file \"$origin\" to destination \"$dest\"
-
+  log " > Linking origin file \"$origin\" to destination \"$dest\""
+  if [[ ! -e "$origin" ]]; then
+    log " >> Origin $origin does not exist" $RED
+    return 1
+  fi
   if [[ -f "$dest" || -d "$dest" ]] && [ ! -L "$dest" ]; then
-      echo "Destination ($dest) already exists. Renaming to $dest-old"
+      log " > Destination ($dest) already exists. Renaming to $dest-old" $YELLOW
       mv "$dest" "$dest-old"
   fi
   ln -sfn "$origin" "$dest"
 }
 
+#-------------------------------------------------------------------------------
+# Script start
+# ------------------------------------------------------------------------------
+
+if [ $(uname -s) == "Darwin" ] && [ ! -d "$SYNC_FOLDER" ]; then
+  log "----------------------------------------------------------------" $REDBOLD
+  log "Could not find the source for private files (Dropbox or GDrive)." $REDBOLD
+  log "Adjust the setup_links.sh script or sync your files first." $REDBOLD
+  log "Your source folder is currently set to $SYNC_FOLDER" $REDBOLD
+  log "----------------------------------------------------------------" $REDBOLD
+  echo ""
+fi
+
+log "Setting links to dotfiles on user home dir: $HOME" $GREEN
+
 # Link .rc files
+log "Linking .rc files" $GREEN
 for FILE in $HOME/.dotfiles/rc/*
 do
   create_link $FILE ~/.$(basename $FILE)
 done
 
 # Link SSH keys
-if [[ ! -d "$HOME/.ssh" ]] && [[ -d "$sync_folder/SSH_Keys" ]]; then
-    ln -sfn "$sync_folder/SSH_Keys" $HOME/.ssh
+log "Linking .ssh directory" $GREEN
+if [[ -d "$SYNC_FOLDER/Configs/SSH_Keys" ]]; then
+    ln -sfn "$SYNC_FOLDER/Configs/SSH_Keys" $HOME/.ssh
 fi
 
 # List of settings to be syncd between computers. Separated by spaces.
-
-# Settings from $HOME/Library/Application Support
-application_support_settings='2fa'
-
 if [ $(uname -s) == "Darwin" ]; then
-  # Link private settings to ~/Library/Application Support
-  for X in $application_support_settings
+  ## Settings from $HOME/Library/Application Support
+  log "Linking ~/Library/Application Support files" $GREEN
+  for X in $(ls "$SYNC_FOLDER/Configs/AppSupport/")
   do
-    if [[ ! -d "$HOME/Library/Application Support/$X" ]]; then
-        create_link "$sync_folder/Configs/$X" "$HOME/Library/Application Support/$X"
-    fi
+    create_link "$SYNC_FOLDER/Configs/AppSupport/$X" "$HOME/Library/Application Support/$X"
   done
 
-  # Link workflows from ~/Library/Services/
-  for X in $(ls "$sync_folder/Configs/automator/")
+  ## Link preferences from ~/Library/Preferences/
+  log "Linking ~/Library/Preferences files" $GREEN
+  for X in $(ls "$SYNC_FOLDER/Configs/Preferences/")
   do
-    create_link "$sync_folder/Configs/automator/"$X "$HOME/Library/Services/$X"
+    create_link "$SYNC_FOLDER/Configs/Preferences/$X" "$HOME/Library/Preferences/$X"
   done
 
-  # Link preferences from ~/Library/Preferences/
-  for X in $(ls "$HOME/.dotfiles/mac/configs/")
+  ## Link workflows from ~/Library/Services/
+  log "Linking ~/Library/Services (automator) files" $GREEN
+  for X in $(ls "$SYNC_FOLDER/Configs/automator/")
   do
-    create_link "$HOME/.dotfiles/mac/configs/$X" "$HOME/Library/Preferences/$X"
+    create_link "$SYNC_FOLDER/Configs/automator/"$X "$HOME/Library/Services/$X"
   done
-
 fi
+
+log "Setting links finished" $GREEN
